# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/interviews:
    get:
      tags:
        - Interviews
      summary: List Interviews
      description: Get a paginated list of Interviews with optional filtering and sorting
      operationId: listInterviews
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Interviews
      summary: Create Interview
      description: Create a new Interview
      operationId: createInterview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInterviewRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/job_applications:
    get:
      tags:
        - JobApplications
      summary: List JobApplications
      description: Get a paginated list of JobApplications with optional filtering and sorting
      operationId: listJobApplications
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobApplicationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - JobApplications
      summary: Create JobApplication
      description: Create a new JobApplication
      operationId: createJobApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobApplicationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobApplication'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Interview:
      type: object
      required:
        - id
        - application_id
        - interviewer_id
        - scheduled_at
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
        interviewer_id:
          type: string
          format: uuid
        scheduled_at:
          type: string
          format: date-time
        round:
          nullable: true
          type: integer
          format: int32
        interview_format:
          nullable: true
          type: string
        rating:
          nullable: true
          type: integer
          format: int32
        feedback:
          nullable: true
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateInterviewRequest:
      type: object
      required:
        - application_id
        - interviewer_id
        - scheduled_at
        - status
        - metadata
      properties:
        application_id:
          type: string
          format: uuid
        interviewer_id:
          type: string
          format: uuid
        scheduled_at:
          type: string
          format: date-time
        round:
          nullable: true
          type: integer
          format: int32
        interview_format:
          nullable: true
          type: string
        rating:
          nullable: true
          type: integer
          format: int32
        feedback:
          nullable: true
          type: string
        status:
          type: string
        metadata:
          type: object
...
    InterviewList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Interview'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    JobApplication:
      type: object
      required:
        - id
        - candidate_id
        - requisition_id
        - stage_id
        - applied_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        candidate_id:
          type: string
          format: uuid
        requisition_id:
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        last_stage_id:
          nullable: true
          type: string
          format: uuid
        stage_updated_at:
          nullable: true
          type: string
          format: date-time
        date_closed:
          nullable: true
          type: string
          format: date-time
        refuse_reason:
          nullable: true
          type: string
        refused_at:
          nullable: true
          type: string
          format: date-time
        applied_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateJobApplicationRequest:
      type: object
      required:
        - candidate_id
        - requisition_id
        - stage_id
        - applied_at
        - metadata
      properties:
        candidate_id:
          type: string
          format: uuid
        requisition_id:
          type: string
          format: uuid
        stage_id:
          type: string
          format: uuid
        last_stage_id:
          nullable: true
          type: string
          format: uuid
        stage_updated_at:
          nullable: true
          type: string
          format: date-time
        date_closed:
          nullable: true
          type: string
          format: date-time
        refuse_reason:
          nullable: true
          type: string
        refused_at:
          nullable: true
          type: string
          format: date-time
        applied_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    JobApplicationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/JobApplication'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []